backend/pyproject.toml
[tool.poetry]
name = "fitness-club-management-system-api"
version = "0.1.0"
description = "FastAPI migration target for Fitness Club Management System"
authors = ["Codex <codex@openai.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.116.1"
uvicorn = {extras = ["standard"], version = "^0.35.0"}
sqlalchemy = "^2.0.43"
asyncpg = "^0.30.0"
alembic = "^1.16.5"
pydantic-settings = "^2.10.1"
python-jose = "^3.5.0"
argon2-cffi = "^25.1.0"
redis = "^6.4.0"
celery = "^5.5.3"
email-validator = "^2.3.0"
python-dateutil = "^2.9.0.post0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-asyncio = "^1.1.0"
pytest-cov = "^7.1.0"
httpx = "^0.28.1"
ruff = "^0.13.0"
aiosqlite = "^0.21.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"
